ocr-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.0.0
numpy==1.26.0
opencv-python-headless==4.8.0.74
pytesseract==0.3.10
imagehash==4.3.1
python-dotenv==1.0.0
